backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --strip-extras requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via starlette
click==8.1.7
    # via uvicorn
dnspython==2.7.0
    # via
    #   email-validator
    #   pymongo
email-validator==2.2.0
    # via pydantic
fastapi==0.115.3
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   email-validator
motor==3.6.0
    # via -r requirements.in
pydantic==2.9.2
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.23.4
    # via pydantic
pydantic-settings==2.6.0
    # via -r requirements.in
pymongo==4.9.2
    # via motor
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
sniffio==1.3.1
    # via anyio
starlette==0.41.0
    # via fastapi
typing-extensions==4.12.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn==0.32.0
    # via -r requirements.in
